liquidity_metasploit/liquidity_metasploit/reinforcement_learning.py
```python
import gym
from gym import spaces
import numpy as np
from .common_vulnerabilities import CommonVulnerabilities
from .attack_tactics import AttackTactics

class ReinforcementLearning:
    def __init__(self):
        self.common_vulnerabilities = CommonVulnerabilities()
        self.attack_tactics = AttackTactics()
        self.state_space = spaces.Discrete(len(self.common_vulnerabilities))
        self.action_space = spaces.Discrete(len(self.attack_tactics))

    def step(self, action):
        attack_tactic = self.attack_tactics[action]
        reward, done = attack_tactic.execute(self.common_vulnerabilities)
        return self.state_space.sample(), reward, done, {}

    def reset(self):
        return self.state_space.sample()

    def render(self):
        pass

    def close(self):
        pass
```